feat(generate): rework example script
